+++ src/server/sv_bot.cpp
@@ -703,6 +703,6 @@ int SV_BotGetSnapshotEntity( int client, int sequence ) {
 	if (sequence < 0 || sequence >= frame->num_entities) {
 		return -1;
 	}
-	return svs.snapshotEntities[(frame->first_entity + sequence) % svs.numSnapshotEntities].number;
+	return svs.snapshotEntities[(frame->first_entity + (int64_t)sequence) % svs.numSnapshotEntities].number;
 }
